+++ src/ui/page_backup/events.rs
@@ -13,20 +13,29 @@ pub fn on_stack_changed() {
 }
 
 pub async fn on_stop_backup_create() -> Result<()> {
-    ui::utils::confirmation_dialog(
-        &gettext("Abort running backup creation?"),
-        &gettext("The backup will remain incomplete if aborted now."),
-        &gettext("Continue"),
-        &gettext("Abort"),
-    )
-    .await?;
+    let operation = BORG_OPERATION.with(|op| Ok::<_, Error>(op.load().active()?.clone()))?;
 
-    BORG_OPERATION.with(|op| {
-        op.load()
-            .active()?
-            .set_instruction(borg::Instruction::Abort(borg::Abort::User));
-        Ok::<_, Error>(())
-    })?;
+    if operation.aborting() {
+        ui::utils::confirmation_dialog(
+            &gettext("Abort Saving Backup State?"),
+            &gettext(
+                "The current backup state is in the process of being saved. The backup can be continued later without saving the state. Some data might have to be copied again.",
+            ),
+            &gettext("Continue"),
+            &gettext("Abort"),
+        )
+        .await?;
+    } else {
+        ui::utils::confirmation_dialog(
+            &gettext("Stop Running Backup?"),
+            &gettext("The current backup state will be saved. You can continue your backup later by starting it again."),
+            &gettext("Continue"),
+            &gettext("Stop"),
+        )
+        .await?;
+    }
+
+    operation.set_instruction(borg::Instruction::Abort(borg::Abort::User));
 
     Ok(())
 }